refactor: move Preprocessor.js -> preprocess dir
