+++ packages/compile/src/index.js
@@ -4,5 +4,5 @@ export { generateCode } from './CodeGen.js'
 export { parseAndGenerate, printNames } from './Entry.js'
 export { canonicalName } from './_shared/helpers.js'
 export { parseModel } from './Parser.js'
-export { preprocessModel } from './Preprocessor.js'
+export { preprocessModel } from './preprocess/preprocessor.js'
 export { readDat } from './ReadDat.js'
